Rename backend-specific files.
Instead of having each backend's files named AdapterBase/PeripheralBase, rename
them with unique names. This is to prepare the codebase for having multiple enabled
backends.
