+++ simpleble/src/backends/android/AdapterAndroid.h
