+++ simpleble/src/backends/android/PeripheralAndroid.h
