+++ simpleble/src/backends/macos/AdapterMac.h
