+++ simpleble/src/backends/macos/PeripheralMac.h
