+++ simpleble/src/backends/plain/PeripheralPlain.h
