+++ simpleble/src/backends/windows/AdapterWindows.h
