+++ simpleble/src/backends/windows/PeripheralWindows.h
